Move execute_mobile_command codes to mixin class
