+++ appium/webdriver/webdriver.py
@@ -41,6 +41,7 @@ from .extensions.clipboard import Clipboard
 from .extensions.context import Context
 from .extensions.device_time import DeviceTime
 from .extensions.execute_driver import ExecuteDriver
+from .extensions.execute_mobile_command import ExecuteMobileCommand
 from .extensions.hw_actions import HardwareActions
 from .extensions.images_comparison import ImagesComparison
 from .extensions.ime import IME
@@ -121,6 +122,7 @@ class WebDriver(
     DeviceTime,
     Display,
     ExecuteDriver,
+    ExecuteMobileCommand,
     Gsm,
     HardwareActions,
     ImagesComparison,
@@ -330,26 +332,6 @@ class WebDriver(
         """
         return MobileWebElement(self, element_id, w3c)
 
-    def press_button(self, button_name):
-        """Sends a physical button name to the device to simulate the user pressing.
-
-        iOS only.
-        Possible button names can be found in
-        https://github.com/appium/WebDriverAgent/blob/master/WebDriverAgentLib/Categories/XCUIDevice%2BFBHelpers.h
-
-        Args:
-            button_name (str): the button name to be sent to the device
-
-        Returns:
-            `appium.webdriver.webdriver.WebDriver`
-
-        """
-        data = {
-            'name': button_name
-        }
-        self.execute_script('mobile: pressButton', data)
-        return self
-
     def set_value(self, element, value):
         """Set the value on an element in the application.
 
@@ -366,29 +348,6 @@ class WebDriver(
         }
         self.execute(Command.SET_IMMEDIATE_VALUE, data)
         return self
-
-    @property
-    def battery_info(self):
-        """Retrieves battery information for the device under test.
-
-        Returns:
-            `dict`: containing the following entries
-                level: Battery level in range [0.0, 1.0], where 1.0 means 100% charge.
-                    Any value lower than 0 means the level cannot be retrieved
-                state: Platform-dependent battery state value.
-                    On iOS (XCUITest):
-                        1: Unplugged
-                        2: Charging
-                        3: Full
-                        Any other value means the state cannot be retrieved
-                    On Android (UIAutomator2):
-                        2: Charging
-                        3: Discharging
-                        4: Not charging
-                        5: Full
-                        Any other value means the state cannot be retrieved
-        """
-        return self.execute_script('mobile: batteryInfo')
 
     # pylint: disable=protected-access
